% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Create a template table for tidying a text-column}
\usage{
fuzzy_match(
  .data,
  stringvar,
  threshold = options("fuzzy_threshold")[[1]],
  .listpool = NULL,
  ...
)
}
\arguments{
\item{.data}{A \code{data.frame} or \code{tbl}}

\item{stringvar}{The name of the column to be tidy (quoted or not)}

\item{threshold}{the minimum distance between strings considered
as OK}

\item{.listpool}{An optional list provided by \code{\link[=fuzzy_pool]{fuzzy_pool()}}}

\item{...}{additional arguments for the function
\code{\link[stringdist:stringdist]{stringdist()}}}
}
\description{
Create a template table for tidying a text-column
}
\examples{
test_df <- data.frame(fruit = c("banana", "blueberry", "limon", "pinapple",
                                "apple", "aple", "Apple", "bonana"))
fuzzy_template <- fuzzy_match(test_df, fruit)
fuzzy_template

}
\seealso{
\code{\link[stringdist:stringdist]{stringdist()}}
}
