% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy.R
\name{fuzzy_pool}
\alias{fuzzy_pool}
\title{Classify strings of characters according to their similarity}
\usage{
fuzzy_pool(string_vec, threshold = options("fuzzy_threshold")[[1]], ...)
}
\arguments{
\item{string_vec}{a vector of strings of characters}

\item{threshold}{the minimum distance between strings considered
as OK}

\item{...}{additional arguments for the function
\code{\link[stringdist:stringdist]{stringdist()}}}
}
\description{
There is no need for the user to call this function directly. This function
is used internally to create a list of strings of characters sorted according
to their similarity.
}
\details{
It is possible to pass to this function arguments for
\code{\link[stringdist:stringdist]{stringdist()}} to fine tune the computation of
the distance between strings.
}
\examples{
test <- c("banana", "blueberry", "APPLE", "apple", "aple", "Apple", "bonana")
fuzzy_pool(test)

}
\seealso{
\code{fuzzy_tidy()}, \code{\link[stringdist:stringdist]{stringdist()}}
}
