% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_colname}
\alias{get_colname}
\title{Extract column name from dataset}
\usage{
get_colname(.data, col)
}
\arguments{
\item{.data}{A \code{data.frame} or \code{tbl}}

\item{col}{A column name}
}
\description{
This function extracts a column name from a dataset irrespective of whether the
column is called by its quoted name, its non quotted name, or through an
object.
}
\examples{
get_colname(iris, "Sepal.Length")
get_colname(iris, Sepal.Length)
get_colname(iris, 1)

}
\seealso{
\code{\link[=get_col]{get_col()}}
}
