% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{merge_common_list}
\alias{merge_common_list}
\title{Merge common elements from a list}
\usage{
merge_common_list(list)
}
\arguments{
\item{list}{A list}
}
\description{
This function merges the commom elements from a list of depth 1.
The implementation is not very efficient, but it does the job for now.
}
\examples{
test <- list(a = c(1, 2), b = c(1, 3), c = 4)
test
merge_common_list(test)

}
\references{
\href{https://stackoverflow.com/questions/47322126/merging-list-with-common-elements}{stackoverflow}
}
